+++ system/controllers/routers.php
@@ -187,5 +187,5 @@ switch ($action) {
         break;
 
     default:
-        echo 'action not defined';
+        r2(U . 'routers/list/', 's', '');
 }